+++ src/midje/internal_ideas/fakes.clj
@@ -22,10 +22,11 @@
                                                    with-altered-roots]]
         [midje.error-handling.exceptions :only [user-error]]
         [midje.internal-ideas.wrapping :only [with-wrapping-target]]
+        [midje.util.deprecation :only [deprecate]]
         [midje.ideas.arrow-symbols]
-        [midje.config :only [*allow-default-prerequisites*]]
         [clojure.tools.macro :only [macrolet]])
-  (:require [clojure.zip :as zip])
+  (:require [clojure.zip :as zip]
+            [midje.config :as config])
   (:import midje.ideas.metaconstants.Metaconstant))
 
 
@@ -165,7 +166,10 @@
 
 
 (defn usable-default-function? [fake]
-  (and *allow-default-prerequisites*
+  (when config/*allow-default-prerequisites*
+    (deprecate "*allow-default-prerequisites* is deprecated and will be removed in Midje 1.6.\nUse config variable :allow-default-prerequisites instead."))
+  (and (or config/*allow-default-prerequisites*
+           (config/choice :allow-default-prerequisites))
        (bound? (:var fake))
        (let [value-in-var (var-get (:var fake))
              unfinished-fun (:midje/unfinished-fun (meta (:var fake)))]